requirements.txt
flask
flask_cors
opencv-python-headless
pillow
